+++ src/Justimmo/Model/Mapper/V1/EmployeeMapper.php
@@ -66,6 +66,9 @@ class EmployeeMapper extends AbstractMapper
             'anrede'         => array(
                 'property' => 'salutation',
             ),
+            'firma'          => array(
+                'property' => 'company',
+            ),
         );
     }